Move LoadVolatile classes into Load.[ch]pp
