+++ jlm/llvm/Makefile.sub
@@ -29,7 +29,6 @@ libllvm_SOURCES = \
     jlm/llvm/ir/operators/GetElementPtr.cpp \
     jlm/llvm/ir/operators/lambda.cpp \
     jlm/llvm/ir/operators/Load.cpp \
-    jlm/llvm/ir/operators/LoadVolatile.cpp \
     jlm/llvm/ir/operators/MemCpy.cpp \
     jlm/llvm/ir/operators/operators.cpp \
     jlm/llvm/ir/operators/Phi.cpp \
@@ -111,7 +110,6 @@ libllvm_HEADERS = \
 	jlm/llvm/ir/operators/operators.hpp \
 	jlm/llvm/ir/operators/Phi.hpp \
 	jlm/llvm/ir/operators/Load.hpp \
-	jlm/llvm/ir/operators/LoadVolatile.hpp \
 	jlm/llvm/ir/operators/MemCpy.hpp \
 	jlm/llvm/ir/operators/gamma.hpp \
 	jlm/llvm/ir/operators/GetElementPtr.hpp \
@@ -155,7 +153,6 @@ libllvm_TESTS += \
 	tests/jlm/llvm/frontend/llvm/test-restructuring \
 	tests/jlm/llvm/frontend/llvm/test-select \
 	tests/jlm/llvm/ir/operators/LoadTests \
-	tests/jlm/llvm/ir/operators/LoadVolatileTests \
 	tests/jlm/llvm/ir/operators/MemCpyTests \
 	tests/jlm/llvm/ir/operators/TestCall \
 	tests/jlm/llvm/ir/operators/test-ConstantFP \
